requirements.txt
flask
flask-cors
pymongo[srv]
pyjwt
bcrypt
python-dotenv
pypdf
reportlab
pytesseract
Pillow
gunicorn